feat(schemas:delta2): split set commands into multiple files based on device-modules
